// src/schema/community/tests/communityCreate/communityCreate.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type CreateCommunityMutationVariables = Types.Exact<{
  input: Types.CreateCommunityInput;
}>;


export type CreateCommunityMutation = { __typename?: 'Mutation', createCommunity: { __typename?: 'Community', id: string, name: string | null, description: string | null, status: Types.CommnunityStatus } };


export const CreateCommunity = gql`
    mutation CreateCommunity($input: CreateCommunityInput!) {
  createCommunity(input: $input) {
    id
    name
    description
    status
  }
}
    `;